+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [To be deprecated]
 - Deprecate options "features" and "feature".
 - Remove "/features" api path.
+- Remove "features" getter from Server
 
 ## [unreleased]
 ### Added